+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## Release 0.3.0
+
+- Fix connection
+- Add disconnect functionality
+- General Refactors
+
+## Release 0.2.1 and 0.2.2
+
+- Query recognizeId in user and return products when login mutation
+
 ## Release 0.2.0
 
 - Add authentication flux